Refactor SSO redirector for customization; add SSO MFA ceremony to redirector.
