+++ tool/tctl/sso/tester/command.go
@@ -35,6 +35,7 @@ import (
 	"github.com/gravitational/teleport/api/utils/keys"
 	"github.com/gravitational/teleport/lib/auth/authclient"
 	"github.com/gravitational/teleport/lib/client"
+	"github.com/gravitational/teleport/lib/client/sso"
 	"github.com/gravitational/teleport/lib/cryptosuites"
 	"github.com/gravitational/teleport/lib/defaults"
 	"github.com/gravitational/teleport/lib/service/servicecfg"
@@ -139,7 +140,7 @@ func (cmd *SSOTestCommand) ssoTestCommand(ctx context.Context, c *authclient.Cli
 		}
 
 		// note: loginErr is processed further down.
-		loginResponse, loginErr := cmd.runSSOLoginFlow(ctx, raw.Kind, c, requestInfo.Config)
+		loginResponse, loginErr := cmd.runSSOLoginFlow(ctx, raw.Kind, c, requestInfo.SSOLoginConsoleRequestFn)
 
 		if requestInfo.RequestCreateErr != nil {
 			return trace.BadParameter("Failed to create auth request. Check the auth connector definition for errors. Error: %v", requestInfo.RequestCreateErr)
@@ -165,15 +166,38 @@ func (cmd *SSOTestCommand) TryRun(ctx context.Context, selectedCommand string, c
 
 // AuthRequestInfo is helper type, useful for tying together test handlers of different auth types.
 type AuthRequestInfo struct {
-	// Config holds *client.RedirectorConfig used for SSO redirect.
-	Config *client.RedirectorConfig
+	// SSOLoginConsoleRequestFn allows customizing issuance of SSOLoginConsoleReq. Optional.
+	SSOLoginConsoleRequestFn func(req client.SSOLoginConsoleReq) (*client.SSOLoginConsoleResponse, error)
 	// RequestID is ID of auth request created for SSO test.
 	RequestID string
 	// RequestCreateErr holds an error in case auth request creation failed.
 	RequestCreateErr error
 }
 
-func (cmd *SSOTestCommand) runSSOLoginFlow(ctx context.Context, protocol string, c *authclient.Client, config *client.RedirectorConfig) (*authclient.SSHLoginResponse, error) {
+// SSOLoginConsoleRequestFn allows customizing issuance of SSOLoginConsoleReq. Optional.
+type SSOLoginConsoleRequestFn func(req client.SSOLoginConsoleReq) (*client.SSOLoginConsoleResponse, error)
+
+func (cmd *SSOTestCommand) runSSOLoginFlow(ctx context.Context, connectorType string, c *authclient.Client, initiateSSOLoginFn SSOLoginConsoleRequestFn) (*authclient.SSHLoginResponse, error) {
+	proxies, err := c.GetProxies()
+	if err != nil {
+		return nil, trace.Wrap(err)
+	}
+
+	if len(proxies) == 0 {
+		return nil, trace.BadParameter("cluster has no proxies.")
+	}
+
+	rdConfig := sso.RedirectorConfig{
+		ProxyAddr: proxies[0].GetPublicAddr(),
+		Browser:   cmd.Browser,
+	}
+
+	rd, err := sso.NewRedirector(ctx, rdConfig)
+	if err != nil {
+		return nil, trace.Wrap(err)
+	}
+	defer rd.Close()
+
 	sshKey, tlsKey, err := cryptosuites.GenerateUserSSHAndTLSKey(ctx, cryptosuites.GetCurrentSuiteFromAuthPreference(c))
 	if err != nil {
 		return nil, trace.Wrap(err)
@@ -187,41 +211,22 @@ func (cmd *SSOTestCommand) runSSOLoginFlow(ctx context.Context, protocol string,
 		return nil, trace.Wrap(err)
 	}
 
-	proxies, err := c.GetProxies()
-	if err != nil {
-		return nil, trace.Wrap(err)
-	}
-
-	if len(proxies) == 0 {
-		return nil, trace.BadParameter("cluster has no proxies.")
-	}
-
-	cfg := client.MakeDefaultConfig()
-	cfg.WebProxyAddr = proxies[0].GetPublicAddr()
-	cfg.Browser = cmd.Browser
-
-	tc, err := client.NewClient(cfg)
-	if err != nil {
-		return nil, trace.Wrap(err)
-	}
-
-	return client.SSHAgentSSOLogin(ctx, client.SSHLoginSSO{
-		SSHLogin: client.SSHLogin{
-			ProxyAddr:         tc.WebProxyAddr,
-			SSHPubKey:         ssh.MarshalAuthorizedKey(sshPub),
-			TLSPubKey:         tlsPub,
-			TTL:               tc.KeyTTL,
-			Insecure:          tc.InsecureSkipVerify,
-			Pool:              nil,
-			Compatibility:     tc.CertificateFormat,
-			RouteToCluster:    tc.SiteName,
-			KubernetesCluster: tc.KubernetesCluster,
-		},
+	req := client.SSOLoginConsoleReq{
 		ConnectorID: "-sso-test",
-		Protocol:    protocol,
-		BindAddr:    tc.BindAddr,
-		Browser:     tc.Browser,
-	}, config)
+		RedirectURL: rd.ClientCallbackURL,
+		SSOUserPublicKeys: client.SSOUserPublicKeys{
+			SSHPubKey: ssh.MarshalAuthorizedKey(sshPub),
+			TLSPubKey: tlsPub,
+		},
+	}
+
+	initResp, err := initiateSSOLoginFn(req)
+	if err != nil {
+		return nil, trace.Wrap(err)
+	}
+
+	resp, err := rd.SSOCeremony(ctx, initResp.RedirectURL)
+	return resp, trace.Wrap(err)
 }
 
 // GetDiagMessage is helper function for preparing message set to be shown to user.
