+++ tool/tctl/sso/tester/github.go
@@ -75,7 +75,7 @@ func githubTest(c *authclient.Client, connector types.GithubConnector) (*AuthReq
 		return &client.SSOLoginConsoleResponse{RedirectURL: request.RedirectURL}, nil
 	}
 
-	requestInfo.Config = &client.RedirectorConfig{SSOLoginConsoleRequestFn: makeRequest}
+	requestInfo.SSOLoginConsoleRequestFn = makeRequest
 	return requestInfo, nil
 }
