+++ tool/tctl/sso/tester/saml.go
@@ -85,7 +85,7 @@ func samlTest(c *authclient.Client, samlConnector types.SAMLConnector) (*AuthReq
 		return &client.SSOLoginConsoleResponse{RedirectURL: request.RedirectURL}, nil
 	}
 
-	requestInfo.Config = &client.RedirectorConfig{SSOLoginConsoleRequestFn: makeRequest}
+	requestInfo.SSOLoginConsoleRequestFn = makeRequest
 	return requestInfo, nil
 }
